\name{cplselection-package}
\alias{cplselection-package}
\alias{cplselection}
\docType{package}
\title{
\packageTitle{cplselection}
}
\description{
\packageDescription{cplselection}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cplselection}
\packageIndices{cplselection}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cplselection}

Maintainer: \packageMaintainer{cplselection}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
