% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplselection.R
\name{cplselection}
\alias{cplselection}
\title{This function generates iterates from Bayesian copula sample selection model}
\usage{
cplselection(
  outcome_formula,
  select_formula,
  outcome_dist,
  select_dist,
  data,
  loop,
  stepsize,
  stepadj,
  sliceadj,
  truevals = NULL,
  truedata = NULL
)
}
\arguments{
\item{outcome_formula}{outcome regression equation}

\item{select_formula}{selection regression equation}

\item{outcome_dist}{outcome variable distribution}

\item{select_dist}{selection variable distribution}

\item{data}{the data matrix with named column variables}

\item{loop}{total MCMC draws}

\item{stepsize}{step size for MH sampling}

\item{stepadj}{step adjustment}

\item{sliceadj}{slicing adjustment}

\item{truevals}{if true process is known, provide true parameters here}

\item{truedata}{if true process is known, provide complete data set here}

\item{burnin}{burn-in count}
}
\description{
This function generates iterates from Bayesian copula sample selection model
}
