% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplselectionInfer.R
\name{cplselectionInfer}
\alias{cplselectionInfer}
\title{Conduct Bayesian inference with given MCMC iterates}
\usage{
cplselectionInfer(iterates, burnin)
}
\arguments{
\item{iterates}{A list of MCMC iterates matrix returned by `cplselectionMCMC` funciton}

\item{burnin}{Burn-in iterations to discard.}
}
\value{
A data.frame object containing estimation results.
}
\description{
Conduct Bayesian inference with given MCMC iterates
}
\examples{
NULL
}
