% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcplEstimate.R
\name{simcplEstimate}
\alias{simcplEstimate}
\title{Simulate sample data and estimate Bayesian copula sample selection model}
\usage{
simcplEstimate(
  seed,
  filedir,
  filename,
  outcome_dist,
  select_dist,
  outcome_par,
  select_par,
  theta,
  x_dist,
  nobs,
  vis,
  outcome_formula,
  select_formula,
  loop,
  burnin,
  stepsize,
  stepadj,
  sliceadj
)
}
\arguments{
\item{seed}{}

\item{filedir}{}

\item{filename}{}

\item{outcome_dist}{}

\item{select_dist}{}

\item{outcome_par}{}

\item{select_par}{}

\item{theta}{}

\item{x_dist}{}

\item{nobs}{}

\item{vis}{visibility}

\item{outcome_formula}{}

\item{select_formula}{}

\item{loop}{}

\item{burnin}{}

\item{stepsize}{}

\item{stepadj}{}

\item{sliceadj}{}
}
\description{
Simulate sample data and estimate Bayesian copula sample selection model
}
\examples{
NULL
}
