% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{This function simulates a data set of sample selection with use-specified parameters.}
\usage{
simdata(
  outcome_dist,
  outcome_par,
  select_dist,
  select_par,
  theta,
  nobs,
  vis,
  x_dist,
  selection = TRUE
)
}
\arguments{
\item{outcome_dist}{Marginal distribution of outcome variable.}

\item{outcome_par}{A list of marginal parameters for the outcome equation. Must include "beta".}

\item{select_dist}{Marginal distribution of selection variable.}

\item{select_par}{A list of marginal parameters for the selection equation. Must include "beta".}

\item{theta}{The dependence parameter in the Gaussian copula function.}

\item{nobs}{Number of observations.}

\item{x_dist}{A list of explanatory variable distributional specification lists. 
Options are "Normal", "Uniform", "Poisson", "Negative Binomial" and "Normal Mixture".}

\item{selection}{Whether sample selection happens.}

\item{x_par}{A list of explanatory variable distributional parameters.}
}
\description{
This function simulates a data set of sample selection with use-specified parameters.
}
