% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_All.R
\name{simstudy_All}
\alias{simstudy_All}
\title{Run all simulation studies with cplselection functions}
\usage{
simstudy_All(seeds, folderpath, rest.time = 0, mc.cores = 8)
}
\arguments{
\item{seeds}{random seeds}

\item{folderpath}{directory to save the .RData results}

\item{rest.time}{if to avoid overheating on personal device, time of resting}

\item{mc.cores}{parallel cores}
}
\description{
Run all simulation studies with cplselection functions
}
\examples{
NULL
}
