% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_Heckit.R
\name{simstudy_Heckit}
\alias{simstudy_Heckit}
\title{Performs simulation study with heckit method}
\usage{
simstudy_Heckit(seeds, folderpath, rest.time)
}
\arguments{
\item{seeds}{}

\item{folderpath}{}

\item{rest.time}{}
}
\description{
Performs simulation study with heckit method
}
