% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_LogitNorm.R
\name{simstudy_LogitNorm}
\alias{simstudy_LogitNorm}
\title{Simulation study : Logit - Normal}
\usage{
simstudy_LogitNorm(seeds, folderpath, rest.time = 600)
}
\arguments{
\item{seeds}{}

\item{folderpath}{}

\item{rest.time}{}
}
\description{
Simulation study : Logit - Normal
}
\examples{
NULL
}
