% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_OLS.R
\name{simstudy_OLS}
\alias{simstudy_OLS}
\title{Performs simulation study with heckit method}
\usage{
simstudy_OLS(seeds, folderpath, rest.time)
}
\arguments{
\item{seeds}{}

\item{folderpath}{}

\item{rest.time}{}
}
\value{
A data.frame object containing results.
}
\description{
Performs simulation study with heckit method
}
\examples{
NULL
}
