% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_performance.R
\name{simstudy_performance}
\alias{simstudy_performance}
\title{Generate Performance Metrics Table for Simulation Studies}
\usage{
simstudy_performance(
  restab,
  by = c("x_dist", "dependence", "selectivity"),
  type = "Bayes",
  trim = 0
)
}
\arguments{
\item{restab}{Result summary table}
}
\value{
A data.frame object containing performance metrics.
}
\description{
Generate Performance Metrics Table for Simulation Studies
}
\examples{
NULL
}
