% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_summary.R
\name{simstudy_summary}
\alias{simstudy_summary}
\title{Summarize the Bayesian estimation results of input files}
\usage{
simstudy_summary(folderpath)
}
\value{
A summary table of Bayesian estimates and true parameters
}
\description{
Summarize the Bayesian estimation results of input files
}
\examples{
none.
}
