% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplselectionInfer.R
\name{cplselectionInfer}
\alias{cplselectionInfer}
\title{Infer estimates from MCMC iterates.}
\usage{
cplselectionInfer(iterates, burnin, trueval = NULL)
}
\arguments{
\item{iterates}{A list of MCMC iterates returned by `cplselectionMCMC` function.}

\item{burnin}{Burn-in iterations to discard.}

\item{trueval}{True parameters, if available. Default set to NULL.}
}
\value{
A data.frame object containing estimation results.
}
\description{
This function makes Bayesian inference with given MCMC iterates returned by `cplselectionMCMC` function.
}
\examples{
NULL
}
