% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplselectionInfer.R
\name{cplselectionInfer}
\alias{cplselectionInfer}
\title{Infer estimates from MCMC iterates.}
\usage{
cplselectionInfer(iterates, burnin, trueval = NULL)
}
\arguments{
\item{iterates}{A list of MCMC iterates returned by `cplselectionMCMC` function.}

\item{burnin}{Burn-in iterations to discard.}

\item{trueval}{True parameters, if available. Default set to NULL.}
}
\value{
A data.frame object containing estimation results.
}
\description{
This function makes Bayesian inference with given MCMC iterates returned by `cplselectionMCMC` function.
}
\examples{
# Example data : Poisson outcome variable YO, logit selection process YS
nobs <- 300
dependence <- 0.5
betaO0 <- 1
betaO1 <- 1
betaS0 <- 0
betaS1 <- 1
betaS2 <- 1
uo <- runif(nobs)
us <- pnorm(dependence*qnorm(uo) + dependence*rnorm(nobs))
data <- data.frame(X1 = rnorm(nobs), X2 = rnorm(nobs))
data$YO <- qpois(p=uo, lambda=exp(betaO0 + betaO1*data$X1))
data$YS <- qbinom(p=us, prob=1/(1+exp(-(betaS0 + betaS1*data$X1 + betaS2*data$X2))), size=1)
data$YO[data$YS==0] <- NA

# Perform MCMC draws
draws <- cplselectionMCMC(outcome_formula = YO~X1, select_formula = YS~X1+X2,
                          outcome_dist = "Poisson", select_dist = "Logit", 
                          data = data, loop = 5000)
                          
# Conduct Inference
# if the true values are not known
result.1 <- cplselectionInfer(iterates=draws, burnin=loop/2) 
print(round(result.1, 2))

# if the true values are known
trueval <- c(dependence, betaO0, betaO1, betaS0, betaS1, betaS2)
result.2 <- cplselectionInfer(iterates=draws, burnin=loop/2, trueval=trueval)
print(round(result.2, 2))

}
