% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplselectionMCMC.R
\name{cplselectionMCMC}
\alias{cplselectionMCMC}
\title{MCMC Sampling of Bayesian copula sample selection model}
\usage{
cplselectionMCMC(
  outcome_formula,
  select_formula,
  outcome_dist,
  select_dist,
  data,
  loop,
  stepsize = 0.1,
  stepadj = 0.05,
  sliceadj = 500
)
}
\arguments{
\item{outcome_formula}{}

\item{select_formula}{}

\item{outcome_dist}{}

\item{select_dist}{}

\item{data}{}

\item{loop}{}

\item{stepsize}{}

\item{stepadj}{}

\item{sliceadj}{}
}
\value{
a list of MCMC draws
}
\description{
MCMC Sampling of Bayesian copula sample selection model
}
\examples{
NULL
}
