% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHeckitEstimate.R
\name{simHeckitEstimate}
\alias{simHeckitEstimate}
\title{Simulate data and estimate with Heckit method}
\usage{
simHeckitEstimate(
  seed,
  outcome_dist,
  select_dist,
  outcome_par,
  select_par,
  theta,
  x_dist,
  nobs,
  outcome_formula,
  select_formula
)
}
\arguments{
\item{seed}{}

\item{outcome_dist}{}

\item{select_dist}{}

\item{outcome_par}{}

\item{select_par}{}

\item{theta}{}

\item{x_dist}{}

\item{nobs}{}

\item{outcome_formula}{}

\item{select_formula}{}
}
\description{
Simulate data and estimate with Heckit method
}
