% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_performance.R
\name{simstudy_performance}
\alias{simstudy_performance}
\title{Generate Performance Metrics Table for Simulation Studies}
\usage{
simstudy_performance(restab, details = TRUE)
}
\arguments{
\item{restab}{Result summary table}
}
\value{
Performance metrics
}
\description{
Generate Performance Metrics Table for Simulation Studies
}
